WIP - Frontend implementation.
Refactored useFileManager.ts into useOrganizationImages.ts.
Refactored MediaImageCarouselxxx.vue files to use useOrganizationImages.ts.

Some cleanup of backend files.
