+++ frontend/components/media/image-carousel/MediaImageCarouselFull.vue
@@ -1,8 +1,6 @@
 <!-- SPDX-License-Identifier: AGPL-3.0-or-later -->
 <template>
   <div class="relative">
-    <!-- MediaImageCarousel is the main image display component. -->
-    <!-- 'fullscreen' just makes the images a bit bigger in the carousel. -->
     <MediaImageCarousel
       @upload-complete="fetchOrganizationImages"
       :fullscreen="false"
@@ -10,13 +8,12 @@
       :imageUrls="imageUrls"
     />
     <button
-      @click="openMediaImageCarousel()"
-      @keydown.enter="openMediaImageCarousel()"
+      @click="openMediaImageCarousel"
+      @keydown.enter="openMediaImageCarousel"
       class="focus-brand absolute right-2 top-2 z-10 hidden rounded-lg border border-black/80 bg-white/80 p-1 text-black/80 dark:border-white/80 dark:bg-black/80 dark:text-white/80 md:block"
     >
       <Icon :name="IconMap.FULL_SCREEN" size="1.5em" />
     </button>
-    <!-- ModalMediaImageCarousel is the full-screen/modal/popup. -->
     <ModalMediaImageCarousel
       @upload-complete="fetchOrganizationImages"
       @closeModal="handleCloseMediaImageCarousel"
@@ -27,21 +24,23 @@
 
 <script setup lang="ts">
 import type { Swiper } from "swiper/types";
-
 import { useModalHandlers } from "~/composables/useModalHandlers";
+import { useOrganizationImages } from "~/composables/useOrganizationImages";
 import { IconMap } from "~/types/icon-map";
+
+const props = defineProps<{ organizationId?: string }>();
 
 const {
   openModal: openMediaImageCarousel,
   handleCloseModal: handleCloseMediaImageCarousel,
 } = useModalHandlers("ModalMediaImage");
+const { imageUrls, fetchOrganizationImages } = useOrganizationImages(
+  props.organizationId
+);
 
-const props = defineProps<{
-  organizationId?: string;
-}>();
+const swiperRef = ref<{ swiper: Swiper | null }>();
 
-onMounted(() => {
-  fetchOrganizationImages();
+function setupSwiper() {
   const swiperEl = document.querySelector("swiper-container");
   if (swiperEl) {
     swiperRef.value = { swiper: swiperEl.swiper };
@@ -49,57 +48,10 @@ onMounted(() => {
       swiperRef.value = { swiper: swiperEl.swiper };
     });
   }
+}
+
+onMounted(() => {
+  fetchOrganizationImages();
+  setupSwiper();
 });
-
-// TODO: All image-related logic could be moved to a composable / store.
-const colorMode = useColorMode();
-const imageColor = colorMode.value;
-
-const defaultImageUrls = [
-  `${GET_ACTIVE_IMAGE_URL}_${imageColor}.png`,
-  `${GET_ORGANIZED_IMAGE_URL}_${imageColor}.png`,
-  `${GROW_ORGANIZATION_IMAGE_URL}_${imageColor}.png`,
-];
-
-const imageUrls = ref(defaultImageUrls);
-const uploadError = ref(false);
-const swiperRef = ref<{ swiper: Swiper | null }>();
-
-interface OrganizationImage {
-  id: string;
-  fileObject: string;
-  creation_date: string;
-}
-
-async function fetchOrganizationImages() {
-  if (props.organizationId) {
-    try {
-      const response = await fetch(
-        `${BASE_BACKEND_URL}/communities/organizations/${props.organizationId}/images/`,
-        {
-          headers: {
-            Authorization: `Token ${localStorage.getItem("accessToken")}`,
-          },
-        }
-      );
-
-      if (response.ok) {
-        const data = await response.json();
-        if (data.length > 0) {
-          imageUrls.value = data.map(
-            (img: OrganizationImage) => img.fileObject
-          );
-          uploadError.value = false;
-        } else {
-          imageUrls.value = defaultImageUrls;
-        }
-      } else {
-        uploadError.value = true;
-      }
-    } catch (error) {
-      console.error("Error fetching organization images:", error);
-      uploadError.value = true;
-    }
-  }
-}
 </script>
