+++ frontend/components/modal/upload-images/ModalUploadImages.vue
@@ -98,7 +98,7 @@ import draggable from "vuedraggable";
 
 import { IconMap } from "~/types/icon-map";
 
-const { files, handleFiles, removeFile, uploadFiles } = useFileManager();
+const { files, handleFiles, removeFile, uploadFiles } = useOrganizationImages();
 
 export interface Props {
   uploadLimit?: number;
